Refactor media upload
