+++ src/features/groups/group-metadata-form/index.tsx
@@ -1,13 +1,15 @@
 import { EyeClosedIcon, EyeIcon, LockIcon, LockOpenIcon } from 'lucide-react';
 import { editGroupMetadata, Nip29GroupMetadata } from 'nostr-hooks/nip29';
-import { useCallback, useState } from 'react';
+import { useCallback, useRef, useState } from 'react';
+
 import { UploadImageButton } from '@/features/chats/chat-bottom-bar/components';
+
 import { Button } from '@/shared/components/ui/button';
 import { Input } from '@/shared/components/ui/input';
 import { Label } from '@/shared/components/ui/label';
 import { Textarea } from '@/shared/components/ui/textarea';
 import { useToast } from '@/shared/components/ui/use-toast';
-import { useGroupMetadataForm } from './hooks';
+import { useUploadMedia } from '@/shared/hooks';
 
 export type GroupMetadataFormProps = {
   relay: string | undefined;
@@ -16,14 +18,18 @@ export type GroupMetadataFormProps = {
 };
 
 export const GroupMetadataForm = ({ initialMetadata, relay, groupId }: GroupMetadataFormProps) => {
+  const groupImageUrlInputRef = useRef<HTMLInputElement>(null);
   const [name, setName] = useState(initialMetadata?.name ?? '');
   const [picture, setPicture] = useState(initialMetadata?.picture ?? '');
   const [about, setAbout] = useState(initialMetadata?.about ?? '');
   const [isPublic, setIsPublic] = useState(initialMetadata?.isPublic ?? true);
   const [isOpen, setIsOpen] = useState(initialMetadata?.isOpen ?? true);
   const { toast } = useToast();
-  const { groupImageUrlInputRef, openUploadImageDialog, isUploadingGroupImage } =
-    useGroupMetadataForm(setPicture);
+  const { isUploadingMedia, openUploadMediaDialog } = useUploadMedia(
+    setPicture,
+    ['image/jpeg', 'image/png', 'image/gif'],
+    groupImageUrlInputRef,
+  );
   const handleSubmit = useCallback(() => {
     relay &&
       groupId &&
@@ -72,8 +78,8 @@ export const GroupMetadataForm = ({ initialMetadata, relay, groupId }: GroupMeta
             />
             <span className="[&_.animate-spin]:h-6 [&_.animate-spin]:w-6 bg-gray-200 dark:bg-gray-900 rounded-md rounded-s-none absolute top-0 end-0">
               <UploadImageButton
-                isUploadingMedia={isUploadingGroupImage}
-                openUploadMediaDialog={openUploadImageDialog}
+                isUploadingMedia={isUploadingMedia}
+                openUploadMediaDialog={openUploadMediaDialog}
               />
             </span>
           </div>
