+++ src/shared/hooks/index.ts
@@ -3,4 +3,5 @@ export { useActiveRelay } from './use-active-relay';
 export { useCopyToClipboard } from './use-copy-to-clipboard';
 export { useLoginModalState } from './use-login-modal-state';
 export { useSendContent } from './use-send-content';
+export { useUploadMedia } from './use-upload-media';
 export { useZapModalState } from './use-zap-modal-state';
